+++ src/buttongroup.v
@@ -45,7 +45,7 @@ pub fn (mut this ButtonGroup[T]) setup() {
 
 			mut tar := e.target
 			if mut tar is Button {
-				e.target.is_selected = true
+				tar.is_selected = true
 			}
 			this.selected = e.target
 			this.invoke_mouse_up(e.target, e.ctx)