+++ src/textfield.v
@@ -1,6 +1,11 @@
 module iui
 
 import gx
+import gg
+
+const (
+	numbers_val = ['0', '1', '2', '3', '4', '5', '6', '7', '8', '9', '.']
+)
 
 // TextField Component - Single line text input
 pub struct TextField {
@@ -159,5 +164,93 @@ fn (mut this TextField) mouse_down_caret(ctx &GraphicsContext) {
 			this.carrot_left = i
 			return
 		}
+	}
+}
+
+fn (mut app Window) runebox_key(key gg.KeyCode, ev &gg.Event, mut com TextField) {
+	if !com.is_selected {
+		return
+	}
+
+	if key == .right {
+		com.carrot_left += 1
+		return
+	} else if key == .left {
+		com.carrot_left -= 1
+		return
+	}
+	mod := ev.modifiers
+	if mod == 8 {
+		// Windows Key
+		return
+	}
+	if mod == 2 {
+		com.ctrl_down = true
+	}
+	if key == .backspace {
+		com.text = com.text.substr_ni(0, com.carrot_left - 1) +
+			com.text.substr_ni(com.carrot_left, com.text.len)
+		com.carrot_left -= 1
+		com.ctrl_down = false
+		return
+	}
+
+	if key == .left_shift || key == .right_shift {
+		app.shift_pressed = true
+		return
+	}
+
+	enter := is_enter(key)
+
+	if enter {
+		com.last_letter = 'enter'
+		bevnt := com.before_txtc_event_fn(mut app, *com)
+		if bevnt || key == .up || key == .down {
+			return
+		}
+		com.text_change_event_fn(app, com)
+		com.ctrl_down = false
+		return
+	}
+
+	if ev.typ == .key_down {
+		return
+	}
+
+	resu := utf32_to_str(ev.char_code)
+	letter := resu
+
+	com.last_letter = letter
+
+	bevnt := com.before_txtc_event_fn(mut app, *com)
+	if bevnt || key == .up || key == .down {
+		// 'true' indicates cancel event
+		return
+	}
+
+	if mod != 2 && !enter {
+		if com.numeric {
+			if letter !in iui.numbers_val {
+				com.last_letter = letter
+				com.text_change_event_fn(app, com)
+				return
+			}
+		}
+
+		com.text = com.text.substr_ni(0, com.carrot_left) + letter +
+			com.text.substr_ni(com.carrot_left, com.text.len)
+
+		com.carrot_left += 1
+	}
+
+	if enter {
+		com.last_letter = 'enter'
+	} else {
+		com.last_letter = letter
+	}
+	com.text_change_event_fn(app, com)
+	com.ctrl_down = false
+	unsafe {
+		resu.free()
 	}
 }